🚚 rename: 공유하는 타입, 훅, 데이터들 shared 폴더로 분리
