+++ nova-app/src/features/profile/types/api.ts
@@ -1,13 +1,11 @@
-/**
- * 멤버 정보 요청 DTO
- */
+import { MemberLevel } from '@/shared/types/memberLevel';
+
+// 멤버 요청 정보 DTO
 export type MemberRequestDto = {
   name: string;
 };
 
-/**
- * 멤버 정보 응답 DTO
- */
+// 멤버 정보 응답 DTO
 export type MemberResponseDto = {
   id: number;
   name: string;
@@ -15,27 +13,13 @@ export type MemberResponseDto = {
   profileImage: string | null;
 };
 
-/**
- * 멤버 정보 수정 응답 DTO
- */
+// 멤버 정보 수정 응답 DTO
 export type MemberUpdateResponseDto = {
   id: number;
   name: string;
 };
 
-/**
- * 멤버 레벨 enum
- */
-export enum MemberLevel {
-  NOVICE = 'NOVICE', // 입문자
-  BEGINNER = 'BEGINNER', // 초급자
-  INTERMEDIATE = 'INTERMEDIATE', // 중급자
-  ADVANCED = 'ADVANCED', // 숙련자
-}
-
-/**
- * 멤버 개인화 설정 DTO
- */
+// 멤버 개인화 설정 DTO
 export type MemberPersonalizationDto = {
   level: MemberLevel | null;
   background: string | null;
@@ -43,12 +27,9 @@ export type MemberPersonalizationDto = {
   keywords: string[];
 };
 
-/**
- * 연결된 계정 조회 응답 DTO
- */
+// 연결된 계정 조회 응답 DTO
 export type MemberConnectedAccountsResponseDto = {
   googleConnected: boolean;
   kakaoConnected: boolean;
   githubConnected: boolean;
 };
-
