+++ nova-app/src/shared/data/PersonalizationText.ts
