+++ nova-app/src/shared/types/memberLevel.ts
@@ -1,0 +1,7 @@
+// 멤버 레벨 enum
+export enum MemberLevel {
+  NOVICE = 'NOVICE', // 입문자
+  BEGINNER = 'BEGINNER', // 초급자
+  INTERMEDIATE = 'INTERMEDIATE', // 중급자
+  ADVANCED = 'ADVANCED', // 숙련자
+}
